Bubble_Trust_Backend/Bubble_Trust_Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
sqlalchemy
python-dotenv